+++ lib/presentation/mixins/init_config_mixin.dart
@@ -33,10 +33,15 @@ mixin InitConfigMixin {
   }
 
   Future<void> initConfigMobile() async {
+    var isConfigured = true;
     try {
       AppConfig.loadEnvironment();
     } catch (e) {
+      isConfigured = false;
       Logs().e('[ConfigLoader] Config mobile error', e);
+    }
+    if (!AppConfig.initConfigCompleter.isCompleted) {
+      AppConfig.initConfigCompleter.complete(isConfigured);
     }
   }
 }